chore: remove unused headway configuration
Nothing currently loads this configuration, and we don't expect to use
it in the near future. It may have been added "looking ahead" to v2
Busway screens, but we are initially not going to implement headway
mode on these, and the final implementation may use different fields
(or no fields at all).