feat: simplify audio configuration
All audio config values (except the periodic readout offset for bus
shelters) are currently identical across all screens of a given type,
and we expect/want this to continue to be true. We can simplify by
making these un-configurable and essentially embedding the config for
each screen type in Screens, as is currently done for e.g. the readout
interval.

This also adds one field, indicating whether to enable periodic audio
readouts. We expect to need this for upcoming bus shelter screens that
will have audio buttons.