chore: remove unused headway configuration
This field is presently unused, and we have no plans to use it in the
near future. Since we can use `signs.json` from `realtime_signs` to
derive a "headway key" from a stop ID, we should not need to configure
these manually.

The DUP-specific `Headway` configuration is left as-is since this is
actively used in the current DUP logic. It may become unused once DUPs
are migrated to the new departures logic.